feat: add ScoreCard component and update evaluation dialog layout
