+++ app/analytics/components/metrics.tsx
